chore: Change CA location <-> FB location
