+++ src/kutokit/model/ProjectXML.java
@@ -10,6 +10,9 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import kutokit.model.cse.ControlAction;
+import kutokit.model.cse.Controller;
+import kutokit.model.cse.Feedback;
 import kutokit.model.utm.UCA;
 import kutokit.view.components.*;
