+++ src/kutokit/model/cse/Feedback.java
@@ -1,71 +1,68 @@
-package kutokit.view.components;
+package kutokit.model.cse;
 
 import java.util.ArrayList;
 
-import kutokit.model.cse.Components;
-import kutokit.view.components.Controller;
-
-public class ControlAction {
+public class Feedback {
 
 	double startX, startY, endX, endY;
 	Controller controller, controlled;
 	Components dataStore;
-	//xml
+	// xml
 	int id;
-	ArrayList<String> CA;
 	public int controllerID, controlledID;
-	
-	public ControlAction() {
-		
+	ArrayList<String> FB;
+
+	public Feedback() {
+
 	}
-	
-	public ControlAction(String controller, String controlled,  ArrayList<String> CA, int id, Components dataStore) {
+
+	public Feedback(String controller, String controlled, ArrayList<String> FB, int id, Components dataStore) {
 		this.controller = dataStore.findController(controller);
 		this.controlled = dataStore.findController(controlled);
 		this.controllerID = this.controller.getId();
 		this.controlledID = this.controlled.getId();
 		this.id = id;
-		this.CA = CA;
+		this.FB = FB;
 	}
-	
+
 	public Controller getController() {
 		return controller;
 	}
-	
+
 	public Controller getControlled() {
 		return controlled;
 	}
-	
+
 	public int getControllerID() {
 		return controllerID;
 	}
-	
+
 	public int getControlledID() {
 		return controlledID;
 	}
-	
+
 	public int getId() {
 		return id;
 	}
-	
-	public ArrayList<String> getCA() {
-		return CA;
+
+	public ArrayList<String> getFB() {
+		return FB;
 	}
-	
+
 	public void setController(String controller) {
 		this.controller = dataStore.findController(controller);
 	}
-	
+
 	public void setControlled(String controlled) {
 		this.controlled = dataStore.findController(controlled);
 	}
-	
+
 	public void setId(int id) {
 		this.id = id;
 	}
-	
-	public void setCA(ArrayList<String> CA) {
-		this.CA = CA;
+
+	public void setFB(ArrayList<String> FB) {
+		this.FB = FB;
 	}
-	
+
 }
