+++ src/kutokit/view/CseController.java
@@ -38,6 +38,9 @@ import javax.xml.bind.Unmarshaller;
 
 import kutokit.MainApp;
 import kutokit.model.cse.Components;
+import kutokit.model.cse.ControlAction;
+import kutokit.model.cse.Controller;
+import kutokit.model.cse.Feedback;
 import kutokit.view.components.*;
 import kutokit.view.popup.*;
 import kutokit.view.popup.cse.AddCAPopUpController;
@@ -247,7 +250,7 @@ public class CseController {
 					case "feedback":
 						AddFBpop = loader.getController();
 						
-						if(AddCApop.OKclose) {
+						if(AddFBpop.OKclose) {
 							Feedback fb = new Feedback(AddFBpop.controlled, AddFBpop.controller, AddFBpop.FB, dataStore.curId, dataStore);
 							
 							DoubleProperty  startX1 = null, startY1 = null, endX1 = null,  endY1 = null;
